Making this a backwards compatible change.
We accomplish this by adding the new client subscriptions in addition to the old default. The client will only run one instance of a check at a time so this should be safe in the case where someone upgrades and still has the legacy subscription. This gives a safe upgrade path. After you upgrade your sensu installs to 0.26+ you should remove the legacy client subscription as it will be removed in the future.

Signed-off-by: Ben Abrams <me@benabrams.it>